requirements.txt
Flask==2.2.2
Pillow==8.4.0
Werkzeug==2.0.1